% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasterChristian.R
\name{ChristianEaster}
\alias{ChristianEaster}
\title{EasterChristian}
\usage{
ChristianEaster(Year)
}
\arguments{
\item{Year}{vector in integer numeric format: yyyy}
}
\value{
List of vector of: (1) vector of months (2) vector of last easter sundays
}
\description{
Determination of the Christian Easter. The following algorithm for
computing the date of Easter is based on the algorithm of Oudin (1940). It is
valid for any Gregorian year.
}
\examples{

library(RMoon)

Easter <- ChristianEaster(2024)

}
\references{
J.-M. Oudin (1940) Étude sur la date de Pâques, Bulletin Astronomique, 12, Paris
}
