% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffDatesMeeus.R
\name{DiffdaysMeeus}
\alias{DiffdaysMeeus}
\title{diffDatesMeeus}
\usage{
DiffdaysMeeus(Mdate1, Mdate2)
}
\arguments{
\item{Mdate1, Mdate2}{Numerical matrix which need 6 columns (years, months,
days, hours, minutes, seconds) and the same number of rows to make the comparison}
}
\value{
A numeric vector giving a difference of days between each row of M1 and
M2 with decimals
}
\description{
This function computes the difference in Julian days, with decimals,
between 2 dates taking into account the restriction of the function
JulianDayMeeus(), this means that difference of days is computed considering
Julian Calendar all data before 1582-10-04 and Gregorian calendar all data after
1582-10-15.
}
\examples{

library(RMoon)

Mdate1 <-matrix(nrow = 2, ncol = 6)
Mdate2 <-matrix(nrow = 2, ncol = 6)

Mdate1[1, ] <- c(1582, 10, 04, 0, 0, 0)
Mdate1[2, ] <- c(2000, 1, 1, 0, 0, 0)

Mdate2[1, ] <- c(1582, 10, 15, 0, 0, 0)
Mdate2[2, ] <- c(2024, 4, 7, 0, 0, 0)

dif_days <- DiffdaysMeeus(Mdate1, Mdate2)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
