% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthVSOP87.R
\docType{data}
\name{earthVSOP87}
\alias{earthVSOP87}
\title{earthVSOP87}
\format{
.rda
}
\usage{
This dataset is used internally in some function of the RMoon package to get precise computations
}
\description{
The Planetary solutions VSOP87 (Variations Seculaires des Orbites Planetaires)
are analytical solutions of the motion of the planets in different versions. The main version
VSOP87 consists of the series in elliptic elements as in the case of VSOP82 solution and the
other versions VSOP87 (A-B-C-D-E) are built in in rectangular and spherical variables.
}
\examples{

library(RMoon)
df = earthVSOP87

}
\references{
VI/81 Planetary Solutions VSOP87 (Bretagnon+, 1988)
}
\keyword{Earth}
\keyword{VSOP87}
\keyword{dataset}
