% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecliptic2Equatorial.R,
%   R/Equatorial2Horizontal.R
\name{Ecliptic2Ecuatorial}
\alias{Ecliptic2Ecuatorial}
\title{Ecliptic2Equatorial}
\usage{
Ecliptic2Ecuatorial(LocalHourAngleT, DeclinationT, LatitudeT)

Ecliptic2Ecuatorial(LocalHourAngleT, DeclinationT, LatitudeT)
}
\arguments{
\item{LocalHourAngleT}{Angle of the Local Hour (in hours)}

\item{DeclinationT}{Declination (in degrees)}

\item{LatitudeT}{The latitud of the observer (in degrees)}

\item{lambdaT}{Ecliptical or celestial longitude, measured from vernal equinnox}

\item{betaT}{Ecliptical or celestial latitude, positive if north, negative if south}

\item{eclipR}{True Obliquity of the ecliptic (corrected by aberration and nutation)}
}
\value{
Equatorial coordinates in a vector of 7 components:
Hour of alpha (or right ascension),
minutes of alpha
seconds of alpha
sign of declination
degrees of delta (or declination)
minutes of delta
seconds of delta

List of 2 vectors of degrees in horizontal coordinates:
Azimuth,
Height
}
\description{
Conversion of ecliptical coordinates to Equatorial

Conversion of Equatorial coordinates to Horizontal
}
\examples{


library(RMoon)
equatorial <- Ecliptic2Equatorial(lambdaT = 133.167265,
                                  betaT = -3.2291226,
                                  eclipR = 23.440636)



library(RMoon)
horizontal <- Equatorial2Horizontal(LocalHourAngleT = 4.29014,
                                  DeclinationT = -6.71989,
                                  LatitudeT = 38.92139)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2

Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
