% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecliptic2Equatorial.R, R/Equatorial2Ecliptic.R
\name{Ecliptic2Equatorial}
\alias{Ecliptic2Equatorial}
\title{Ecliptic2Equatorial}
\usage{
Ecliptic2Equatorial(lambdaT, betaT, eclipT)

Ecliptic2Equatorial(lambdaT, betaT, eclipT)
}
\arguments{
\item{lambdaT}{Ecliptical or celestial longitude, measured from vernal equinnox}

\item{betaT}{Ecliptical or celestial latitude, positive if north, negative if south}

\item{eclipT}{True Obliquity of the ecliptic (corrected by aberration and nutation)}

\item{RA}{Right Ascension (in hours)}

\item{Dec}{Declination (in degrees)}
}
\value{
Equatorial coordinates in a list of 2 vectors:
RA or right ascension (in hours),
Declination (in degrees)

Ecliptical coordinates in a list of 2 vectors:
lambda (longitude in degrees),
beta (latitude in degrees)
}
\description{
Conversion of ecliptical coordinates to Equatorial

Conversion of Equatorial coordinates to Ecliptical
}
\examples{


library(RMoon)
equatorial <- Ecliptic2Equatorial(lambdaT = 133.162655,
                                  betaT = -3.229126,
                                  eclipT = 23.440636)


library(RMoon)
ecliptic <- Equatorial2Ecliptic(RA = 8.8932,
                                Dec = 19.182,
                                eclipT = 23.440636)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1

Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
