% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equatorial2Ecliptic.R
\name{Equatorial2Ecliptic}
\alias{Equatorial2Ecliptic}
\title{Equatorial2Ecliptic}
\usage{
Equatorial2Ecliptic(RA, Dec, eclipT)
}
\arguments{
\item{RA}{Right Ascension (in hours)}

\item{Dec}{Declination (in degrees)}

\item{eclipT}{True Obliquity of the ecliptic (corrected by aberration and nutation)}
}
\value{
Ecliptical coordinates in a list of 2 vectors:
lambda (longitude in degrees),
beta (latitude in degrees)
}
\description{
Conversion of Equatorial coordinates to Ecliptical
}
\examples{

library(RMoon)
ecliptic <- Equatorial2Ecliptic(RA = 8.8932,
                                Dec = 19.182,
                                eclipT = 23.440636)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
