% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equatorial2Horizontal.R
\name{Equatorial2Horizontal}
\alias{Equatorial2Horizontal}
\title{Equatorial2Horizontal}
\usage{
Equatorial2Horizontal(LocalHourAngleT, DeclinationT, LatitudeT)
}
\arguments{
\item{LocalHourAngleT}{Angle of the Local Hour (in hours)}

\item{DeclinationT}{Declination (in degrees)}

\item{LatitudeT}{The latitud of the observer (in degrees)}
}
\value{
List of 2 vectors of degrees in horizontal coordinates:
Azimuth,
Height
}
\description{
Conversion of Equatorial coordinates to Horizontal
}
\examples{


library(RMoon)
horizontal <- Equatorial2Horizontal(LocalHourAngleT = 4.29014,
                                  DeclinationT = -6.71989,
                                  LatitudeT = 38.92139)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
