% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horizontal2Equatorial.R
\name{Horizontal2Equatorial}
\alias{Horizontal2Equatorial}
\title{Horizontal2Equatorial}
\usage{
Horizontal2Equatorial(AzimuthT, HeightT, LatitudeT)
}
\arguments{
\item{AzimuthT}{Azimuth (in degrees)}

\item{HeightT}{Height (in degrees)}

\item{LatitudeT}{The latitude of the observer (in degrees)}
}
\value{
List of 2 vectors:
LocalHourAngle (in degrees)
Declination (in degrees)
}
\description{
Conversion of Horizontal coordinates to Equatorial
}
\examples{

library(RMoon)
equatorial <- Horizontal2Equatorial(
  AzimuthT = 185.9468,
  HeightT = 44.18548,
  LatitudeT = 38.92139)

}
\references{
Based on Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
