% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JulianDayMeeus.R
\name{JulianDayMeeus}
\alias{JulianDayMeeus}
\title{JulianDayMeeus}
\usage{
JulianDayMeeus(year, month, day, hour, minute, second)
}
\arguments{
\item{year, month, day, hour, minute, seconds}{Numerical vectors that take the temporal
reference of one or several days in UTM time}
}
\value{
List of 3 vectors containing the following elements:
(1) Julian day (2)  Julian century (3) Julian millenium
}
\description{
Provides the Julian Day using 5 parameter from year up to minutes
(seconds can be considered using decimal in the last parameter). This function has
been tested for BCE dates and takes into account dates previous to 05-10-1582 which
are in Julian calendar. Note that this function gives a non-sense day between
05-10-1582 and 14-10-1582 which where drop from the Gregorian calendar. Only if
it is required 15-10-1582, 10 days must be rested to the result.
}
\examples{

library(RMoon)
jlData<- JulianDayMeeus(year = 1992, month = 4, day = 12, hour = 0, minute = 0, second = 0)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
