% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliantoGregorianDate.R
\name{JuliantoGregorianDate}
\alias{JuliantoGregorianDate}
\title{JuliantoGregorianDate}
\usage{
JuliantoGregorianDate(julianday)
}
\arguments{
\item{julianday}{Numerical vectors that take the temporal
reference of one or several days in UTM time}
}
\value{
List of 6 vectors containing the following elements in Gregorian calendar:
(1) year (2)  month (3) day (4) hour (5) minute (6) second
}
\description{
Provides the date in year, month, day, hour, minute and second from a
positive julian day. Only this is valid for positive julian day according to the chapter
7 of Astronomical Algorithms
}
\examples{

library(RMoon)
DatesYMDhms <- JuliantoGregorianDate(julianday = 2436116.31)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
