% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonBrightLimb.R
\name{MoonBrightLimb}
\alias{MoonBrightLimb}
\title{MoonBrightLimb}
\usage{
MoonBrightLimb(alpham, deltam, alphas, deltas)
}
\arguments{
\item{alpham}{Moon's right ascension in degrees}

\item{deltam}{Moon's declination in degrees}

\item{alphas}{Sun's right ascension in degrees}

\item{deltas}{Moon's declination in degrees}
}
\value{
Vector of the position angle of the Moon's bright limb of the midpoint
of the illuminated limb of the moon
}
\description{
Computes the fraction of Moon surface illuminated by the Sun
according to Chapter 48 of Astronomical Algorithms
}
\examples{

#alpham, deltam, distm, alphas, deltas, dists were computed at 12.04.1992, 00h 00m
library(RMoon)
MoonBrightLimb <- MoonBrighLimb(alpham  = 134.6885,
                               deltam  = 13.7684,
                               alphas  = 20.6579,
                               deltas = 8.6964)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
