% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonEclipsesDates.R
\docType{data}
\name{MoonEclipsesDates}
\alias{MoonEclipsesDates}
\title{MoonEclipsesDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Lunar Eclipses since 3000 BC up to 3000 AC
}
\examples{

library(RMoon)
df = MoonEclipsesDates

}
\references{
Dates based on "Eclipse Predictions by Fred Espenak, www.EclipseWise.com"
}
\keyword{dataset}
