% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonFirstQDates.R
\docType{data}
\name{MoonFirstQDates}
\alias{MoonFirstQDates}
\title{MoonFirstQDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Lunar First Quarter since 2000 BC up to 4000 AC
}
\examples{

library(RMoon)
df = MoonFirstQDates

}
\references{
Dates based on "Moon Phases Table courtesy of Fred Espenak, www.Astropixels.com"
}
\keyword{dataset}
