% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonFullDates.R
\docType{data}
\name{MoonFullDates}
\alias{MoonFullDates}
\title{MoonFullQDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Full Lunar since 2000 BC up to 4000 AC with Lunar eclipses information
}
\examples{

library(RMoon)
df = MoonFullDates

}
\references{
Dates based on "Moon Phases Table courtesy of Fred Espenak, www.Astropixels.com"
}
\keyword{dataset}
