% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonGeograficData.R
\docType{data}
\name{MoonGeographicData}
\alias{MoonGeographicData}
\title{MoonGeographicData}
\format{
.rda
}
\usage{
This dataset offer latitude and longitude data of visible and non visible face of the Moon
}
\description{
Moon geography of craters, monts, etc provide by the NASA
}
\examples{

library(RMoon)
df = MoonGeographicData

}
\references{
Dates extracted from NASA website https://www.nasa.gov
}
\keyword{dataset}
