% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonGeoLat.R
\name{MoonGeoLat}
\alias{MoonGeoLat}
\title{MoonGeoLat}
\usage{
MoonGeoLat(L, D, M, M_, F_, A1, A3, E, E_2)
}
\arguments{
\item{L}{Elemental parameter: Moon's mean longitude}

\item{D}{Elemental parameter: Mean eleongation of the Moon}

\item{M}{Elemental parameter: Sun's mean anomaly}

\item{M_}{Elemental parameter: Moon's mean anomaly}

\item{F_}{Elemental parameter: Moon's argument of latitude}

\item{A1}{Elemental parameter: Adjustments by Venus perturbation}

\item{A3}{Elemental parameter: Adjustments by Jupiter perturbation}

\item{E}{Elemental parameter: Adjustments by Earth excentricity}

\item{E_2}{Elemental parameter: Squared Adjustments by Earth excentricity}
}
\value{
Geocentric latitude
}
\description{
Provides the Geocentric Latitude of the Moon
according to Chapter 45 of Astronomical Algorithms
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
GeoLat<- MoonGeoLat(L  = 134.29018627,
                    D  = 113.84230856,
                    M  = 97.643513610,
                    M_ = 5.1508393327,
                    F_ = 219.88972622,
                    A1 = 109.56847763,
                    A3 = 229.53,
                    E  = 1.0001942441,
                    E_2 = 1.0001942441**2)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
