% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonGeoLongDist.R
\name{MoonGeoLongDist}
\alias{MoonGeoLongDist}
\title{MoonGeoLongDist}
\usage{
MoonGeoLongDist(L, D, M, M_, F_, A1, A2, E)
}
\arguments{
\item{L}{Elemental parameter: Moon's mean longitude}

\item{D}{Elemental parameter: Mean eleongation of the Moon}

\item{M}{Elemental parameter: Sun's mean anomaly}

\item{M_}{Elemental parameter: Moon's mean anomaly}

\item{F_}{Elemental parameter: Moon's argument of latitude}

\item{A1}{Elemental parameter: Adjustments by Venus perturbation}

\item{A2}{Elemental parameter: Adjustments by Jupiter perturbation}

\item{E}{Elemental parameter: Adjustments by Earth excentricity}
}
\value{
List of geocentric longitudes of the Moon, distance between the centers
of the Moon and Earth, and equatorial parallax
}
\description{
Provides the Geocentric Longitude and Distance of the Moon
according to Chapter 47 of Astronomical Algorithms
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
GeoLongDist<- MoonGeoLongDist(L  = 134.29018627,
                              D  = 113.84230856,
                              M  = 97.643513610,
                              M_ = 5.1508393327,
                              F_ = 219.88972622,
                              A1 = 109.56847763,
                              A2 = 123.78322533,
                              E  = 1.0001942441)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
