% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonIluminated.R
\name{MoonIlluminated}
\alias{MoonIlluminated}
\title{MoonIlluminated}
\usage{
MoonIlluminated(alpham, deltam, distm, alphas, deltas, dists)
}
\arguments{
\item{alpham}{Moon's right ascension in degrees}

\item{deltam}{Moon's declination in degrees}

\item{distm}{Distance Earth - Moon in Km}

\item{alphas}{Sun's right ascension in degrees}

\item{deltas}{Moon's declination in degrees}

\item{dists}{Distance Earth - Sun in AUs}
}
\value{
Vector of percentage of illuminated Moon's surface according to position of
the Moon and Sun in a determinated date
}
\description{
Computes the fraction of Moon surface illuminated by the Sun
according to Chapter 48 of Astronomical Algorithms
}
\examples{

#alpham, deltam, distm, alphas, deltas, dists were computed at 12.04.1992, 00h 00m
library(RMoon)
MoonSurfaceIlluminated <- MoonIlluminated(alpham  = 134.6885,
                                          deltam  = 13.7684,
                                          distm = 368410,
                                          alphas  = 20.6579,
                                          deltas = 8.6964,
                                          dists = 1.0024977)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
