% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonLastQDates.R
\docType{data}
\name{MoonLastQDates}
\alias{MoonLastQDates}
\title{MoonLastQDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Lunar Last Quarters since 2000 BC up to 4000 AC
}
\examples{

library(RMoon)
df = MoonLastQDates

}
\references{
Dates based on "Moon Phases Table courtesy of Fred Espenak, www.Astropixels.com"
}
\keyword{dataset}
