% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonNewDates.R
\docType{data}
\name{MoonNewDates}
\alias{MoonNewDates}
\title{MoonNewDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Neomenias since 2000 BC up to 4000 AC with Solar eclipses information
}
\examples{

library(RMoon)
df = MoonNewDates

}
\references{
Dates based on "Moon Phases Table courtesy of Fred Espenak, www.Astropixels.com"
}
\keyword{dataset}
