% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonNodes.R
\name{MoonNodes}
\alias{MoonNodes}
\title{MoonNodes}
\usage{
MoonNodes(day, node)
}
\arguments{
\item{day}{Date in format: "yyyy-mm-dd"}

\item{node}{Valid values: "ascending", "descending"}
}
\value{
List of vector of: dates after date given when the Moon reach the nearest
next chosen node.
If not valid value or another values are given at node parameter, the resultant
position will be referred for the following ascending node
}
\description{
Determination of the instant when the Moon Eclipse reaches one of
its nodes according to Chapter 51 of Astronomical Algorithms
}
\examples{

library(RMoon)

MoonEclipseElements <- MoonNodes("1987-05-15", "ascending")

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
