% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonPeriApogeo.R
\name{MoonPeriApogeo}
\alias{MoonPeriApogeo}
\title{MoonPeriApogeo}
\usage{
MoonPeriApogeo(year, month, day)
}
\arguments{
\item{month}{Vector of months in UT}

\item{day}{Vector of days in UT}

\item{decyear}{Vector of years in UT}
}
\value{
List of 2 vectors: Date when next perigeo-apogeo happens and equatorial parallax associated to them
}
\description{
Calculation about when the distance between Moon and Earth is a minimum
(perigeo) or a maximum (apogeo). This function has to be used at current day at gregorian
calendar
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
MoonPeriApogeo(year = c(1988),
                month = c(10),
                day = c(1)

}
\references{
Function created based on algorithms from Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
