% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonPhasePrediction.R
\name{MoonPhasePrediction}
\alias{MoonPhasePrediction}
\title{MoonPhasePrediction}
\usage{
MoonPhasePrediction(day, phase)
}
\arguments{
\item{day}{Date in format: "yyyy-mm-dd"}

\item{phase}{Choice of the Moon's phase you have to choose between: "new",
"full", "firstquarter", "lastquarter"}
}
\value{
List of vector of julian days and dates at UTC
}
\description{
Prediction of the principal Moon's phases: new, full, first quarter,
last quarter, according to Chapter 49 of Astronomical Algorithms
}
\examples{

library(RMoon)

MoonPhase <- MoonPhasePrediction("1977-02-15", "new")

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
