% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuslimtoGregorianDate.R
\name{MuslimtoGregorian}
\alias{MuslimtoGregorian}
\title{MuslimtoGregorianDate}
\usage{
MuslimtoGregorian(MuslimY, MuslimM, MuslimD)
}
\arguments{
\item{MuslimY}{Numerical value for the year of a Muslin year}

\item{MuslimM}{Numerical value for the month of a Muslin month}

\item{MuslimD}{Numerical value for the day of a Muslin day}
}
\value{
List of 5 vectors containing the following elements in Gregorian calendar:
(1) year (2)  month (3) day (4) julian day (5) Information about leap or normal year
}
\description{
Convert a Muslim date in a Gregorian date. This conversion has sense for
Muslim date greater or equal a year = 1, month = 1 and day = 1 which is equivalent
to 16 - 07 - 622. This function offers the julian day and gives information about
if the Muslim year used is a leap or normal year
}
\examples{

library(RMoon)
ListofGregorianDates <- MuslimtoGregorian(c(1,1421), c(1, 1), c(1, 1))

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
