% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObliquityEcliptic.R
\name{ObliqutiyEcliptic}
\alias{ObliqutiyEcliptic}
\title{ObliquityEcliptic}
\usage{
ObliqutiyEcliptic(juliancent, omega)
}
\arguments{
\item{omega}{Numerical vecto which represents the Sun's true geometric longitude}

\item{julianCent}{Numerical vector which represents a Julian converted in millenium}
}
\value{
Epsilon (Obliquity of the Ecliptic), parameter needed for the Sun's Declination and Right Ascension
}
\description{
Estimation of the obliquity of ecliptic
}
\examples{

#JulianMil is computed at 12.04.1992, 00h 00m
library(RMoon)
MeanLong<- ObliqutiyEcliptic(juliancent = -0.077221081451, omega = 274.39607359)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
