% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrbitMeanLongEarth.R
\name{OrbitMeanLongEarth}
\alias{OrbitMeanLongEarth}
\title{OrbitMeanLongEarth}
\usage{
OrbitMeanLongEarth(julianmil)
}
\arguments{
\item{julianmil}{Numerical vector which represents a Julian converted in millenium}
}
\value{
Mean longitud of the Earth in its orbit at that date
}
\description{
Estimation of the mean longitude of Earth's orbit according to the VSOP87 theory
}
\examples{

#JulianMil is computed at 12.04.1992, 00h 00m
library(RMoon)
MeanLong<- OrbitMeanLongEarth(julianmil = -0.0077221081451)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
