% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodicLongDist.R
\docType{data}
\name{PeriodicLongDist}
\alias{PeriodicLongDist}
\title{PeriodicLongDist}
\format{
.rda
}
\usage{
This dataset is used internally in some function of the RMoon package to get precise computations
}
\description{
Main periodic terms for the longitude and distance of the munto. The
unit is 0.000001 degree for longitude and 0.001 km por distance
}
\examples{

library(RMoon)
df = PeriodicLongDist

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
\keyword{dataset}
