% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodicNutObliq.R
\docType{data}
\name{PeriodicNutObliq}
\alias{PeriodicNutObliq}
\title{PeriodicNutObliq}
\format{
.rda
}
\usage{
This dataset is used internally in some function of the RMoon package to get precise computations
}
\description{
Main periodic terms for the nutation and obliquity at units of
0.0001''
}
\examples{

library(RMoon)
df = PeriodicNutObliq

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
\keyword{dataset}
