% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolarEclipsesDates.R
\docType{data}
\name{SolarEclipsesDates}
\alias{SolarEclipsesDates}
\title{SolarEclipsesDates}
\format{
.rda
}
\usage{
This dataset can allow historic analysis year 0000 at original datasets here is -0001
}
\description{
Dates of Solar Eclipses since 3000 BC up to 3000 AC
}
\examples{

library(RMoon)
df = SolarEclipsesDates

}
\references{
Dates based on "Eclipse Predictions by Fred Espenak, www.EclipseWise.com"
}
\keyword{dataset}
