% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunBasicElements.R
\name{SunBasicElements}
\alias{SunBasicElements}
\title{SunBasicElements}
\usage{
SunBasicElements(juliancent)
}
\arguments{
\item{juliancent}{Numerical vector which represents the Julian Day in centuries
with sufficient number of decimals (recommended at least 9)}
}
\value{
List of Theta (Sun's true longitude), Omega (Sun's true geometric longitude),
lambda (Sun's apparent longitude) and R (Sun's radiovector)
}
\description{
Provides some basic elements of the Sun position required in other
functios for computing the Sun Position according to Chapter 24 of Astronomical Algorithms
}
\examples{

library(RMoon)
Sunbasics <- SunBasicElements(juliancent = -0.077221081451)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
