% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunDeclination01.R
\name{SunDeclination01}
\alias{SunDeclination01}
\title{SunDeclination01}
\usage{
SunDeclination01(epsilon, lambda)
}
\arguments{
\item{epsilon}{Numerical vector which represents the Obliquity of the Ecliptic}

\item{lambda}{Numerical vector which represents the Sun's apparent longitude}
}
\value{
List of Grades, Minutes and Seconds of apparent declinations of the Sun
}
\description{
Provides the Declination of the Sun in the version low
accuracy according to Chapter 24 of Astronomical Algorithms
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
SuncoordDec<- SunDeclination01(epsilon = 23.440491532, lambda = 22.33957501)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
