% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunRadiovectorEarth.R
\name{SunRadioVectorEarth}
\alias{SunRadioVectorEarth}
\title{SunRadioVectorEarth}
\usage{
SunRadioVectorEarth(julianmil)
}
\arguments{
\item{julianmil}{Numerical vector which represents a Julian converted in millenium}
}
\value{
Distant Sun - Earth at that date
}
\description{
Estimation of the radio vector in UAs from Earth to Sun according to the VSOP87 theory
}
\examples{

#JulianMil is computed at 12.04.1992, 00h 00m
library(RMoon)
MeanLong<- SunRadioVectorEarth(julianmil = -0.0077221081451)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
