% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunRightAscention01.R
\name{SunRightAscention01}
\alias{SunRightAscention01}
\title{SunRightAscention01}
\usage{
SunRightAscention01(epsilon, lambda)
}
\arguments{
\item{epsilon}{Numerical vector which represents the eccentricty of the
Earth's orbit}

\item{lambda}{Numerical vector which represents the apparent longitud of the Sun,
referred to the true equinnox of the date}
}
\value{
Vector of Apparent Right Ascention of the Sun in Hour / Minutes /
Seconds
}
\description{
Provides the Right Ascention of the Sun in the version low
accurracy according to Chapter 24 of Astronomical Algorithms
}
\examples{
#epsilon and lambda were computed at 12.04.1992, 00h 00m
Suncoord<- SunRightAscention01(epsilon = 23.440491532, lambda = 22.33957501)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
