% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueObliquityEcliptic.R
\name{TrueObliqutiyEcliptic}
\alias{TrueObliqutiyEcliptic}
\title{TrueObliqutiyEcliptic}
\usage{
TrueObliqutiyEcliptic(juliancent)
}
\arguments{
\item{julianCent}{Numerical vector which represents a Julian converted in millenium}
}
\value{
Vector of epsilons (Obliquity of the Ecliptic) in degrees, parameter needed for the Sun's Declination and Right Ascension
}
\description{
Estimation of the obliquity of ecliptic according to
Chapter 22 of Astronomical Algorithms
}
\examples{

#juliancent is computed at 12.04.1992, 00h 00m
library(RMoon)
TrueObliquity <- TrueObliqutiyEcliptic(juliancent = -0.077221081451)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
