% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecliptic2Ecuatorial.R
\name{Ecliptic2Ecuatorial}
\alias{Ecliptic2Ecuatorial}
\title{Ecliptic2Ecuatorial}
\usage{
Ecliptic2Ecuatorial(lambdaT, betaT, eclipR)
}
\arguments{
\item{lambdaT}{Ecliptical or celestial longitude, measured from vernal equinnox}

\item{betaT}{Ecliptical or celestial latitude, positive if north, negative if south}

\item{eclipR}{True Obliquity of the ecliptic (corrected by aberration and nutation)}
}
\value{
Ecuatorial coordinates in a vector of 7 components:
hours of alpha (or right ascencion),
minutes of alpha
seconds of alpha
sign of declination
degrees of delta (or declination)
minutes of delta
seconds of delta
}
\description{
Conversion of eclitical coordinates to Ecuatorial
}
\examples{


library(RMoon)
ecuatorial <- Ecliptic2Ecuatorial(lambda = 133.167265,
                                  beta = -3.2291226,
                                  epsilon = 23.440636)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
