% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecliptic2Equatorial.R
\name{Ecliptic2Equatorial}
\alias{Ecliptic2Equatorial}
\title{Ecliptic2Equatorial}
\usage{
Ecliptic2Equatorial(lambdaT, betaT, eclipR)
}
\arguments{
\item{lambdaT}{Ecliptical or celestial longitude, measured from vernal equinnox}

\item{betaT}{Ecliptical or celestial latitude, positive if north, negative if south}

\item{eclipR}{True Obliquity of the ecliptic (corrected by aberration and nutation)}
}
\value{
Equatorial coordinates in a vector of 7 components:
Hour of alpha (or right ascension),
minutes of alphA
seconds of alpha
sign of declination
degrees of delta (or declination)
minutes of delta
seconds of delta
}
\description{
Conversion of ecliptical coordinates to Equatorial
}
\examples{


library(RMoon)
equatorial <- Ecliptic2Equatorial(lambdaT = 133.167265,
                                  betaT = -3.2291226,
                                  eclipR = 23.440636)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
