% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecliptic2Equatorial.R
\name{Ecliptic2Equatorial}
\alias{Ecliptic2Equatorial}
\title{Ecliptic2Equatorial}
\usage{
Ecliptic2Equatorial(lambdaT, betaT, eclipT)
}
\arguments{
\item{lambdaT}{Ecliptical or celestial longitude, measured from vernal equinnox}

\item{betaT}{Ecliptical or celestial latitude, positive if north, negative if south}

\item{eclipT}{True Obliquity of the ecliptic (corrected by aberration and nutation)}
}
\value{
Equatorial coordinates in a list of 2 vectors:
RA or right ascension (in hours),
Declination (in degrees)
}
\description{
Conversion of ecliptical coordinates to Equatorial
}
\examples{


library(RMoon)
equatorial <- Ecliptic2Equatorial(lambdaT = 133.162655,
                                  betaT = -3.229126,
                                  eclipT = 23.440636)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
