% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JulianDay.R
\name{JulianDay}
\alias{JulianDay}
\title{JulianDay}
\usage{
JulianDay(year, month, day, hour, minute, second)
}
\arguments{
\item{year, month, day, hour, minute}{Numerical vectors that take the temporal
reference of one or several days in UTM time}
}
\value{
List of 3 vectors containing the following elements:
the Julian time, the Julian day (epoch J2000.0), the Julian century and the Julian Millenium
}
\description{
Provides the Julian Day using 5 parameter from year up to minutes
(seconds can be considered using decimal in the last parameter)
}
\examples{

library(RMoon)
jlData<- JulianDay(year = 1992, month = 4, day = 12, hour = 0, minute = 0, second = 0)

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
