% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonEclipses.R
\name{MoonEclipses}
\alias{MoonEclipses}
\title{MoonEcplipses}
\usage{
MoonEclipses(day)
}
\arguments{
\item{day}{Date in format: "yyyy-mm-dd"}
}
\value{
List of vector of: juliano or Julian day of the following full moon and
when the eclipse would happen, predDate or the date at format yyyy-mm-dd hh-mm -ss
gamma or distance from the centre of the Moon to the axis of the Earth'shadow
in units of the Earth's equatorial raidius, umbraMagn or magnitud of the umbra (if
this value is negative not eclipse at umbra happen), penumbraMag or magnitud of the
penumbra, semiDurationPartialumbra or duration in minutes of partial eclipse at
the umbra phase if this phase happens, semiDurationTotalalumbra or duration in minutes of the
total umbra phase if this phase happens and semiDurationPartialpenumbra or duration
in minutes of the penumbra phase if this phase happens.

Semiduration offers non-valid or NaN values if eclipse in umbra - penumbra does not happen
}
\description{
Determination of the posibility of a Moon eclipse when happen
a full moon phase to Chapter 54 of Astronomical Algorithms
}
\examples{

library(RMoon)

MoonEclipseElements <- MoonEclipses("1977-02-15")

}
\references{
Jean Meeus (1991), Astronomical Algorithms, ISBN 0-943396-35-2
}
