% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonEclipses.R
\name{MoonEclipses}
\alias{MoonEclipses}
\title{MoonEcplipses}
\usage{
MoonEclipses(day)
}
\arguments{
\item{day}{Date in format: "yyyy-mm-dd"}
}
\value{
List of 12 vectors of: (1) juliano or Julian day of the following full moon and
when the eclipse would happen (2) predDate or the date at format yyyy-mm-dd hh-mm -ss
 (3) isEclipse or Moon's eclipse condition (4) nearNode or closest node where the eclipse takes
place (5) gamma or least distance from the centre of the Moon to the axis of the Earth'shadow
in units of the Earth's equatorial raidius (6) u radius of the Moon's umbra cone
(7) rho o penumbra radius (if this has sense),
 (8) sigma or umbra radius (if this has sense) (9) umbraMagn or magnitud of the umbra (if
this value is negative not eclipse at umbra happen) (10) penumbraMag or magnitud of the
penumbra (11) semiDurationPartialumbra or duration in minutes of partial eclipse at
the umbra phase if this phase happens (12) semiDurationTotalalumbra or duration in minutes of the
total umbra phase if this phase happens (13) semiDurationPartialpenumbra or duration
in minutes of the penumbra phase if this phase happens.

Semiduration offers non-valid or NaN values if eclipse in umbra - penumbra does not happen
}
\description{
Determination of the posibility of a Moon Eclipse when happen
a full moon phase according to Chapter 54 of Astronomical Algorithms
}
\examples{

library(RMoon)

MoonEclipseElements <- MoonEclipses("1977-02-15")

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
