% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonPhasesImages.R
\name{MoonPhasesImages}
\alias{MoonPhasesImages}
\title{MoonPhasesImages}
\usage{
MoonPhasesImages(phase, download = F, path = "", size = 400)
}
\arguments{
\item{download}{boolean value F or T}

\item{path}{if download = T, it points where the file will be stored}

\item{size}{by default = 400, it can be changed by the user}

\item{phases}{Numerical vector from 1 to 58, two moon phases by day aprox}
}
\value{
If download = F, the function return the image, if download = T the
function downloads a image pointed by path
}
\description{
This allows to manage 58 images taken during a complete Lunar cicle
taken between 11-03-2024 and 09-04-2024. For a 29.5 lunar cicle, these images shows,
the moon with an interval of 12 hours
}
\examples{

Image <- MoonPhasesImages(7,
                          download = F,
                          path = "",
                          size = 400)

}
\references{
https://svs.gsfc.nasa.gov/gallery/moonphase
}
