% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoonSkyPosition.R
\name{MoonSkyPosition}
\alias{MoonSkyPosition}
\title{MoonSkyPosition}
\usage{
MoonSkyPosition(year, month, day, hour, minute, second, Long, Lat)
}
\arguments{
\item{year}{Vector of years in UT}

\item{month}{Vector of months in UT}

\item{day}{Vector of days in UT}

\item{hour}{Vector of hours in UT}

\item{minute}{Vector of minutes in UT}

\item{second}{Vector of seconds in UT}

\item{Long}{Longitude in degrees of the observer}

\item{Lat}{Latitud in degrees of the observer}
}
\value{
List of 3 vectors: Hour angle, Azimuth and Height in degrees
}
\description{
Given a day and a geographical position, computes the Moon's
horizontal coordinates: Azimuth - Heigh plus Hour angle using the library
the Swiss tables provide by the library swephR
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
MoonSkyPosition(year = c(2024),
                month = c(9),
                day = c(27),
                hour = c(2),
                minute = c(0),
                second = c(0),
                Long = -3 + 37/60 + 40/3600,
                Lat = 40 + 28/60 + 42/3600 )

}
\references{
Function created based on algorithms from Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
