% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutationLong.R
\name{NutationLong}
\alias{NutationLong}
\title{NutationLong}
\usage{
NutationLong(juliancent)
}
\arguments{
\item{julianCent}{Numerical vector which represents a Julian converted in millenium}
}
\value{
Vector of varNutation (Nutation in longitude) in degrees, parameter needed for the Moon's Declination and Right Ascension
}
\description{
Estimation of the longitude of nutation according to
Chapter 22 of Astronomical Algorithms
}
\examples{

#juliancent is computed at 12.04.1992, 00h 00m
library(RMoon)
varNutation <- NutationLong(juliancent = -0.077221081451)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
