% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunRightAscension01.R
\name{SunRightAscension01}
\alias{SunRightAscension01}
\title{SunRightAscension01}
\usage{
SunRightAscension01(epsilon, lambda)
}
\arguments{
\item{epsilon}{Numerical vector which represents the obliquity of the
ecliptic}

\item{lambda}{Numerical vector which represents the apparent longitude of the Sun,
referred to the true equinox of the date}
}
\value{
Vector of apparent Sun' right Ascensions
}
\description{
Provides the Right Ascension of the Sun in the version low
accuracy according to Chapter 25 of Astronomical Algorithms
}
\examples{

#epsilon and lambda were computed at 12.04.1992, 00h 00m
library(RMoon)
SuncoordAsc<- SunRightAscension01(epsilon = 23.440491532, lambda = 22.33957501)

}
\references{
Jean Meeus (1998), Astronomical Algorithms, 2nd Ed, ISBN 0-943396-61-1
}
